fix&feat: create logout btn & fix navigation
